Added better help behaviour
While using this in a different library, I set the "h" flag. It was unexpected to me that it
caused it to crash because the short help flag was overridden, so while this seemed a good idea
while first implementing it, I decided against it and switched to only checking for the help
message if no matches were found.